requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ollama==0.2.0
onnxruntime==1.17.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.0
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.26.1
PyAudio==0.2.14
pygame==2.5.2
regex==2024.5.10
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tiktoken==0.6.0
torch==2.3.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
urllib3==2.2.1
